+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ### Added
 
 ### Changed
+- [MAJOR] Changed `onClick` to `onClicked` in `SingleLineInput`
+- [MINOR] Added `onFrameClicked` to `SingleLineInput`
+- [MINOR] Added `iconRight` and `iconLeft` to `InputFrame`
+- [MINOR] Added `isScrollableVertically` and `isScrollableHorizontally` and `shouldMatchAnchorWidth` and `shouldMatchAnchorHeight` to `Portal`
 
 ### Removed